GH-238 polish tests
